Update dependencies, fix tests
