+++ package.json
@@ -50,6 +50,7 @@
     "prettier:fix": "prettier ./src --write",
     "fake-data": "node ./scripts/fakeData.js",
     "test": "TZ=utc vitest",
+    "test:debug": "TZ=utc vitest --disable-console-intercept --reporter=json",
     "test:ci": "TZ=utc vitest run",
     "test:coverage": "TZ=utc vitest run --coverage"
   },
@@ -61,29 +62,27 @@
     "dist"
   ],
   "dependencies": {
-    "@nuxt/kit": "^3.10.2",
-    "@tusbar/cache-control": "^1.0.1"
+    "@nuxt/kit": "^3.12.2",
+    "@tusbar/cache-control": "^1.0.2"
   },
   "devDependencies": {
     "@faker-js/faker": "^8.4.1",
-    "@nuxt/module-builder": "^0.5.5",
-    "@nuxt/schema": "^3.10.2",
-    "@nuxt/test-utils": "^3.11.0",
+    "@nuxt/module-builder": "^0.7.1",
+    "@nuxt/schema": "^3.12.2",
+    "@nuxt/test-utils": "^3.13.1",
     "@nuxtjs/eslint-config-typescript": "^12.1.0",
-    "@types/node": "^20.11.17",
-    "@vitest/coverage-v8": "^1.2.2",
-    "@vue/test-utils": "^2.4.4",
-    "eslint": "^8.56.0",
+    "@types/node": "^20.14.2",
+    "@vue/test-utils": "^2.4.6",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-vue": "^9.21.1",
-    "happy-dom": "^13.3.8",
-    "nuxt": "^3.10.2",
-    "playwright": "^1.41.2",
-    "prettier": "^3.2.5",
-    "vitepress": "^1.0.0-rc.4",
-    "vitest": "^1.2.2",
-    "vitest-environment-nuxt": "^1.0.0",
-    "vue": "^3.4.19"
+    "eslint-plugin-vue": "^9.26.0",
+    "happy-dom": "^14.12.0",
+    "nuxt": "^3.12.2",
+    "playwright-core": "^1.44.1",
+    "prettier": "^3.3.2",
+    "vitepress": "^1.2.3",
+    "vitest": "^1.6.0",
+    "vue-json-pretty": "^2.4.0"
   }
 }
