+++ src/runtime/serverHandler/helpers/storage.ts
@@ -3,21 +3,7 @@ import type { NuxtMultiCacheSSRContext } from './../../types'
 import { useRuntimeConfig } from '#imports'
 import serverOptions from '#multi-cache-server-options'
 
-const runtimeConfig = useRuntimeConfig()
-
-const cacheContext: NuxtMultiCacheSSRContext = {}
-
-// Initialize all enabled caches. Explicit initialization because some
-// caches might need additional configuration options and/or checks.
-if (runtimeConfig.multiCache.component) {
-  cacheContext.component = createStorage(serverOptions.component?.storage)
-}
-if (runtimeConfig.multiCache.data) {
-  cacheContext.data = createStorage(serverOptions.data?.storage)
-}
-if (runtimeConfig.multiCache.route) {
-  cacheContext.route = createStorage(serverOptions.route?.storage)
-}
+let cacheContext: NuxtMultiCacheSSRContext | null = null
 
 /**
  * Method to initialize the caches.
@@ -26,5 +12,22 @@ if (runtimeConfig.multiCache.route) {
  * afterwards.
  */
 export function loadCacheContext() {
+  if (!cacheContext) {
+    const runtimeConfig = useRuntimeConfig()
+
+    cacheContext = {}
+
+    // Initialize all enabled caches. Explicit initialization because some
+    // caches might need additional configuration options and/or checks.
+    if (runtimeConfig.multiCache.component) {
+      cacheContext.component = createStorage(serverOptions.component?.storage)
+    }
+    if (runtimeConfig.multiCache.data) {
+      cacheContext.data = createStorage(serverOptions.data?.storage)
+    }
+    if (runtimeConfig.multiCache.route) {
+      cacheContext.route = createStorage(serverOptions.route?.storage)
+    }
+  }
   return cacheContext
 }
