+++ test/cdnHeaders.e2e.spec.ts
@@ -1,4 +1,4 @@
-import { fileURLToPath } from 'node:url'
+import path from 'path'
 import { setup, fetch } from '@nuxt/test-utils/e2e'
 import { describe, expect, test, vi } from 'vitest'
 import type { NuxtMultiCacheOptions } from '../src/runtime/types'
@@ -14,41 +14,39 @@ const useRuntimeConfig = vi.fn(() => ({
 
 vi.stubGlobal('useRuntimeConfig', useRuntimeConfig)
 
-describe('The CDN headers feature', async () => {
-  const multiCache: NuxtMultiCacheOptions = {
-    component: {
-      enabled: true,
-    },
-    data: {
-      enabled: true,
-    },
-    route: {
-      enabled: true,
-    },
-    cdn: {
-      enabled: true,
-    },
-    api: {
-      enabled: true,
-      authorization: false,
-      cacheTagInvalidationDelay: 5000,
-    },
-  }
-  const nuxtConfig: any = {
-    multiCache,
-  }
+const multiCache: NuxtMultiCacheOptions = {
+  component: {
+    enabled: true,
+  },
+  data: {
+    enabled: true,
+  },
+  route: {
+    enabled: true,
+  },
+  cdn: {
+    enabled: true,
+  },
+  api: {
+    enabled: true,
+    authorization: false,
+    cacheTagInvalidationDelay: 5000,
+  },
+}
+const nuxtConfig: any = {
+  multiCache,
+}
+await setup({
+  server: true,
+  logLevel: 0,
+  runner: 'vitest',
+  build: true,
+  // browser: true,
+  rootDir: path.resolve(__dirname, './../playground'),
+  nuxtConfig,
+})
 
-  const baseUrl = import.meta.url
-  await setup({
-    server: true,
-    logLevel: 0,
-    runner: 'vitest',
-    build: true,
-    // browser: true,
-    rootDir: fileURLToPath(new URL('../playground', baseUrl)),
-    nuxtConfig,
-  })
-
+describe('The CDN headers feature', () => {
   test('Sets the correct CDN headers', async () => {
     const response = await fetch('/api/cdnHeaders')
     expect(response.headers.get('surrogate-control')).toMatchInlineSnapshot(
