+++ test/composables/useRouteCache.nuxt.spec.ts
@@ -1,14 +1,42 @@
+import { mockNuxtImport } from '@nuxt/test-utils/runtime'
 import { describe, expect, test, vi } from 'vitest'
-import { NuxtMultiCacheCDNHelper } from './../../src/runtime/helpers/CDNHelper'
-import { useCDNHeaders } from './../../src/runtime/composables'
+import { useRouteCache } from './../../src/runtime/composables'
+import { NuxtMultiCacheRouteCacheHelper } from './../../src/runtime/helpers/RouteCacheHelper'
 
-vi.mock('vue', () => {
+mockNuxtImport('useRuntimeConfig', () => {
+  return () => {
+    return {
+      multiCache: {
+        component: true,
+      },
+    }
+  }
+})
+
+vi.mock('vue', async (importOriginal) => {
+  const actual = await importOriginal()
+  const storage: Record<string, any> = {
+    foobar: 'Cached data.',
+  }
   return {
+    // @ts-ignore
+    ...actual,
     useSSRContext: () => {
       return {
         event: {
           context: {
-            __MULTI_CACHE_CDN: new NuxtMultiCacheCDNHelper(),
+            __MULTI_CACHE: {
+              data: {
+                getItem: (key: string) => {
+                  return Promise.resolve(storage[key])
+                },
+                setItem: (key: string, data: any) => {
+                  storage[key] = data
+                  return Promise.resolve()
+                },
+              },
+            },
+            __MULTI_CACHE_ROUTE: new NuxtMultiCacheRouteCacheHelper(),
           },
         },
       }
@@ -19,19 +47,7 @@ vi.mock('vue', () => {
   }
 })
 
-vi.mock('#imports', () => {
-  return {
-    useRuntimeConfig: () => {
-      return {
-        multiCache: {
-          component: true,
-        },
-      }
-    },
-  }
-})
-
-describe('useCDNHeaders composable', () => {
+describe('useRouteCache composable', () => {
   test('Does not call callback in client', () => {
     process.client = true
     const params = {
@@ -39,7 +55,7 @@ describe('useCDNHeaders composable', () => {
     }
 
     const spyCallback = vi.spyOn(params, 'cb')
-    useCDNHeaders(spyCallback as any)
+    useRouteCache(spyCallback as any)
     expect(spyCallback).not.toHaveBeenCalled()
   })
 
@@ -50,7 +66,7 @@ describe('useCDNHeaders composable', () => {
     }
 
     const spyCallback = vi.spyOn(params, 'cb')
-    useCDNHeaders(spyCallback as any)
+    useRouteCache(spyCallback as any)
     expect(spyCallback).toHaveBeenCalledOnce()
   })
 
@@ -58,13 +74,13 @@ describe('useCDNHeaders composable', () => {
     process.client = false
     const dummyHelper = 'dummy helper'
 
-    useCDNHeaders(
+    useRouteCache(
       (helper) => {
         expect(helper).toEqual(dummyHelper)
       },
       {
         context: {
-          __MULTI_CACHE_CDN: dummyHelper,
+          __MULTI_CACHE_ROUTE: dummyHelper,
         },
       } as any,
     )
@@ -73,8 +89,8 @@ describe('useCDNHeaders composable', () => {
   test('Gets the event from SSR context.', () => {
     process.client = false
 
-    useCDNHeaders((helper) => {
-      expect(helper).toHaveProperty('_control')
+    useRouteCache((helper) => {
+      expect(helper).toHaveProperty('tags')
     })
   })
 
@@ -89,11 +105,11 @@ describe('useCDNHeaders composable', () => {
     }
 
     const spyCallback = vi.spyOn(params, 'cb')
-    useCDNHeaders(spyCallback as any)
+    useRouteCache(spyCallback as any)
     expect(spyCallback).not.toHaveBeenCalled()
   })
 
-  test('Does not call callback if CDN helper is missing.', async () => {
+  test('Does not call callback if route helper is missing.', async () => {
     process.client = false
 
     const vue = await import('vue')
@@ -106,7 +122,7 @@ describe('useCDNHeaders composable', () => {
     }
 
     const spyCallback = vi.spyOn(params, 'cb')
-    useCDNHeaders(spyCallback as any)
+    useRouteCache(spyCallback as any)
     expect(spyCallback).not.toHaveBeenCalled()
   })
 })
