--- test/customStorageDriver.test.ts
@@ -1,52 +1,0 @@
-import { fileURLToPath } from 'node:url'
-import { setup, $fetch } from '@nuxt/test-utils/e2e'
-import { describe, expect, test } from 'vitest'
-import type { NuxtMultiCacheOptions } from '../src/runtime/types'
-
-describe('The data cache feature', async () => {
-  const multiCache: NuxtMultiCacheOptions = {
-    component: {
-      enabled: true,
-    },
-    data: {
-      enabled: true,
-    },
-    route: {
-      enabled: true,
-    },
-    cdn: {
-      enabled: true,
-    },
-    api: {
-      enabled: true,
-      authorization: false,
-      cacheTagInvalidationDelay: 5000,
-    },
-  }
-  const nuxtConfig: any = {
-    multiCache,
-  }
-
-  const baseUrl = import.meta.url
-  await setup({
-    server: true,
-    logLevel: 0,
-    runner: 'vitest',
-    build: true,
-    // browser: true,
-    rootDir: fileURLToPath(new URL('../playground', baseUrl)),
-    nuxtConfig,
-  })
-
-  test('Uses a custom storage driver.', async () => {
-    const data = await $fetch('/api/customStorageDriver', {
-      method: 'get',
-    })
-
-    expect(data).toMatchInlineSnapshot(`
-      {
-        "value": "just_an_example_value",
-      }
-    `)
-  })
-})
