+++ test/helpers/CDNHelper.nuxt.spec.ts
