+++ test/helpers/RouteCacheHelper.nuxt.spec.ts
