+++ test/module.nuxt.spec.ts
