+++ test/routeCache.e2e.spec.ts
@@ -1,45 +1,43 @@
-import { fileURLToPath } from 'node:url'
+import path from 'node:path'
 import { setup, $fetch } from '@nuxt/test-utils/e2e'
 import { describe, expect, test } from 'vitest'
 import type { NuxtMultiCacheOptions } from '../src/runtime/types'
-import purgeAll from './__helpers__/purgeAll'
 import { decodeRouteCacheItem } from '../src/runtime/helpers/cacheItem'
+import purgeAll from './__helpers__/purgeAll'
 
-describe('The route cache feature', async () => {
-  const multiCache: NuxtMultiCacheOptions = {
-    component: {
-      enabled: true,
-    },
-    data: {
-      enabled: true,
-    },
-    route: {
-      enabled: true,
-    },
-    cdn: {
-      enabled: true,
-    },
-    api: {
-      enabled: true,
-      authorization: false,
-      cacheTagInvalidationDelay: 5000,
-    },
-  }
-  const nuxtConfig: any = {
-    multiCache,
-  }
+const multiCache: NuxtMultiCacheOptions = {
+  component: {
+    enabled: true,
+  },
+  data: {
+    enabled: true,
+  },
+  route: {
+    enabled: true,
+  },
+  cdn: {
+    enabled: true,
+  },
+  api: {
+    enabled: true,
+    authorization: false,
+    cacheTagInvalidationDelay: 5000,
+  },
+}
+const nuxtConfig: any = {
+  multiCache,
+}
+await setup({
+  server: true,
+  logLevel: 0,
+  runner: 'vitest',
+  build: true,
+  // browser: true,
+  rootDir: path.resolve(__dirname, './../playground'),
+  nuxtConfig,
+})
 
-  const baseUrl = import.meta.url
-  await setup({
-    server: true,
-    logLevel: 0,
-    runner: 'vitest',
-    build: true,
-    // browser: true,
-    rootDir: fileURLToPath(new URL('../playground', baseUrl)),
-    nuxtConfig,
-  })
-
+describe('The route cache feature', () => {
   test('caches a page', async () => {
     await purgeAll()
