+++ test/serverHandler/cdnHeaders.nuxt.spec.ts
