+++ test/settings/index.nuxt.spec.ts
