--- vitest.config.ts
@@ -1,16 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-  test: {
-    globals: true,
-    environment: 'happy-dom',
-    // include: ['test/**/*.test.ts'],
-    coverage: {
-      all: true,
-      provider: 'v8',
-      reporter: ['text', 'json', 'html'],
-      include: ['src/**/*.*'],
-      exclude: ['src/runtime/types.ts'],
-    },
-  },
-})
